backend/requirements.txt
fastapi
uvicorn
openai
pillow
pydantic
python-dotenv
httpx
